Extract POCO mapping into ClickHouse.Driver.Poco namespace
Move the POCO-only types (PocoMappings, PocoTypeRegistry, and the
property-info struct) out of ClickHouse.Driver.Copy — which historically
held bulk-copy machinery — into a dedicated ClickHouse.Driver.Poco
namespace and folder. Bulk-copy serializers stay in Copy/. Also rename
the BinaryInsertPropertyInfo type to PocoPropertyInfo now that it's
shared by both the insert and read paths and lives under Poco/, so the
old "BinaryInsert" prefix would be misleading.

Internal-only refactor; no public API change.
